+++ tencentcloud/cdn/v20180606/models.go
@@ -3397,6 +3397,13 @@ type PurgeUrlsCacheRequest struct {
 
 	// URL 列表，需要包含协议头部 http:// 或 https://
 	Urls []*string `json:"Urls,omitempty" name:"Urls" list`
+
+	// 刷新区域
+	// 无此参数时，默认刷新加速域名所在加速区域
+	// 填充 mainland 时，仅刷新中国境内加速节点上缓存内容
+	// 填充 overseas 时，仅刷新中国境外加速节点上缓存内容
+	// 指定刷新区域时，需要与域名加速区域匹配
+	Area *string `json:"Area,omitempty" name:"Area"`
 }
 
 func (r *PurgeUrlsCacheRequest) ToJsonString() string {